% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{complete_standard_calendar}
\alias{complete_standard_calendar}
\title{Title}
\usage{
complete_standard_calendar(.data, x)
}
\arguments{
\item{.data}{DBI object}

\item{x}{ti object}
}
\value{
DBI object
}
\description{
Title
}
\keyword{internal}
