% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{detect_sql_dialect}
\alias{detect_sql_dialect}
\title{Detect SQL dialect from a DBI connection}
\usage{
detect_sql_dialect(.con)
}
\arguments{
\item{.con}{A DBI connection object}
}
\value{
A character string: "snowflake", "duckdb", or "postgres"
}
\description{
Detect SQL dialect from a DBI connection
}
\keyword{internal}
