% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ff_bob2tif}
\alias{ff_bob2tif}
\title{Convert a brick of bytes to a TIFF image}
\usage{
ff_bob2tif(path, dims, out = fs::path_ext_set(path, "tif"))
}
\arguments{
\item{path}{Input image path.}

\item{dims}{Dimension vector for the TIF image dimensions.}

\item{out}{Output image path. Defaults to the input image path, with .tif extension.}
}
\value{
The image array, invisibly.
}
\description{
Convert a brick of bytes to a TIFF image
}
