% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{ff_compress}
\alias{ff_compress}
\title{Compress a TIFF image}
\usage{
ff_compress(path, type = "deflate", out = path)
}
\arguments{
\item{path}{Input image path.}

\item{type}{Type of compression to apply (\code{"none"}, \code{"LZW"}, \code{"PackBits"}, \code{"RLE"}, \code{"JPEG"}, \code{"deflate"} (default) or \code{"Zip"}).}

\item{out}{Output image path. Defaults to the input image path.}
}
\value{
The output image, invisibly.
}
\description{
Compress a TIFF image
}
