% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{ff_extract}
\alias{ff_extract}
\title{Extract specific frames from a multi-frame TIFF}
\usage{
ff_extract(path, frames, out = fs::path_ext_set(path, "\%frame.tif"))
}
\arguments{
\item{path}{Path of the multi-frame TIFF.}

\item{frames}{Frame numbers to extract.}

\item{out}{Path for the output single-frame images, using "\%frame" as placeholder for the frame number. Defaults to the input file path with extension ".001.tif" etc.}
}
\description{
Extract specific frames from a multi-frame TIFF
}
