% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{ff_invert}
\alias{ff_invert}
\title{Invert a TIFF image}
\usage{
ff_invert(path, out = path)
}
\arguments{
\item{path}{Input image path.}

\item{out}{Output image path. Defaults to the input image path.}
}
\value{
The output image, invisibly.
}
\description{
Invert a TIFF image
}
