% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{ff_merge}
\alias{ff_merge}
\title{Merge single-frame images into a multi-frame TIFF}
\usage{
ff_merge(paths, out)
}
\arguments{
\item{paths}{Paths to the single-frame images to be merged.}

\item{out}{Path to the output multi-frame TIFF.}
}
\description{
Merge single-frame images into a multi-frame TIFF
}
