% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ff_npy2tif}
\alias{ff_npy2tif}
\title{Convert a NumPy array to a TIFF image}
\usage{
ff_npy2tif(path, out = fs::path_ext_set(path, "tif"))
}
\arguments{
\item{path}{Input image path.}

\item{out}{Output image path. Defaults to the input image path, with .tif extension.}
}
\value{
The image array, invisibly.
}
\description{
Convert a NumPy array to a TIFF image
}
