% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{ff_split}
\alias{ff_split}
\title{Split a multi-frame TIFF into single-frame images}
\usage{
ff_split(path, out = fs::path_ext_set(path, "\%frame.tif"))
}
\arguments{
\item{path}{Path of the multi-frame TIFF.}

\item{out}{Path for the output single-frame images, using "\%frame" as placeholder for the frame number. Defaults to the input file path with extension ".001.tif" etc.}
}
\description{
Split a multi-frame TIFF into single-frame images
}
