% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{ff_stretch}
\alias{ff_stretch}
\title{Stretch the contrast of a TIFF image}
\usage{
ff_stretch(path, out = path)
}
\arguments{
\item{path}{Input image path.}

\item{out}{Output image path. Defaults to the input image path.}
}
\value{
The output image, invisibly.
}
\description{
Stretch the contrast of a TIFF image
}
