% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ff_tif2csv}
\alias{ff_tif2csv}
\title{Convert a TIFF image to a CSV file}
\usage{
ff_tif2csv(path, out = fs::path_ext_set(path, "csv"))
}
\arguments{
\item{path}{Input image path.}

\item{out}{Output image path. Defaults to the input image path, with .tif extension.}
}
\value{
The image array, invisibly.
}
\description{
Convert a TIFF image to a CSV file
}
