% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{ff_transpose}
\alias{ff_transpose}
\title{Transpose a TIFF image}
\usage{
ff_transpose(path, dims, out = path)
}
\arguments{
\item{path}{Input image path.}

\item{dims}{The dimension permutation vector.}

\item{out}{Output image path. Defaults to the input image path.}
}
\value{
The output image, invisibly.
}
\description{
Transpose a TIFF image
}
