% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ff_write}
\alias{ff_write}
\title{Write a TIFF image}
\usage{
ff_write(img, path, compression = "deflate")
}
\arguments{
\item{img}{2D or 3D array.}

\item{path}{Path to write to.}

\item{compression}{Type of compression to apply (\code{"none"}, \code{"LZW"}, \code{"PackBits"}, \code{"RLE"}, \code{"JPEG"}, \code{"deflate"} (default) or \code{"Zip"}).}
}
\description{
Write a TIFF image
}
