% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{print_calendar_info}
\alias{print_calendar_info}
\title{Print calendar information block}
\usage{
print_calendar_info(x)
}
\arguments{
\item{x}{ti or segment obj}
}
\value{
print
}
\description{
Print calendar information block
}
\keyword{internal}
