% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-misc.R
\name{sql_date_add}
\alias{sql_date_add}
\title{Generate dialect-appropriate date addition SQL}
\usage{
sql_date_add(.con, unit, n_expr, date_col = "date")
}
\arguments{
\item{.con}{A DBI connection object}

\item{unit}{A character string: "month", "week", etc.}

\item{n_expr}{An expression string for the number of units}

\item{date_col}{A character string for the date column name}
}
\value{
A \code{dplyr::sql} object
}
\description{
Generate dialect-appropriate date addition SQL
}
\keyword{internal}
